Update vacancy announcement page layout
