+++ cfgov/unprocessed/css/pages/careers.less
@@ -20,3 +20,32 @@
         margin: 0 0 0 unit( 5px / @base-font-size-px, em );
     } );
 }
+
+.job-description {
+    .js .o-expandable__read-more[data-js-hook]:not(.o-expandable__expanded) {
+        .respond-to-max( @bp-xs-max, {
+            .o-expandable_content {
+                // shows approx. 10 lines of content
+                height:  unit( @base-line-height-px * 10 / @base-font-size-px, em );
+                overflow: hidden;
+            }
+        });
+    }
+    .embed-container {
+        position: relative; 
+        padding-bottom: 56.25%; 
+        height: 0; 
+        overflow: hidden; 
+        max-width: 100%;
+
+        iframe,
+        object,
+        embed {
+            position: absolute;
+            top: 0;
+            left: 0;
+            width: 100%;
+            height: 100%;
+        }
+    } 
+}
